Add camera image to selected images, add ImagePicker, add picker to ItemEditor
