--- Chronicle/Utility/CameraPicker.swift
@@ -1,52 +1,0 @@
-//
-//  CameraPicker.swift
-//  Chronicle
-//
-//  Created by Skylar Clemens on 9/25/24.
-//
-
-import Foundation
-import UIKit
-import SwiftUI
-
-struct CameraPicker: UIViewControllerRepresentable {
-    @Binding var isPresented: Bool
-    @Binding var image: UIImage?
-    
-    func makeCoordinator() -> Coordinator {
-        Coordinator(parent: self)
-    }
-    
-    func makeUIViewController(context: Context) -> UIImagePickerController {
-        let picker = UIImagePickerController()
-        picker.delegate = context.coordinator
-        picker.sourceType = .camera
-        picker.allowsEditing = true
-        return picker
-    }
-    
-    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) { }
-    
-    class Coordinator: NSObject, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-        var parent: CameraPicker
-        
-        init(parent: CameraPicker) {
-            self.parent = parent
-        }
-        
-        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]) {
-            let editedImage = info[.editedImage] as? UIImage
-            let originalImage = info[.originalImage] as? UIImage
-            
-            if let currentImage = editedImage ?? originalImage {
-                parent.image = currentImage
-            }
-            
-            parent.isPresented = false
-        }
-        
-        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
-            parent.isPresented = false
-        }
-    }
-}
